pyproject.toml
[project]
name = "twinmind-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.0",
    "beautifulsoup4>=4.14.3",
    "google-generativeai>=0.8.5",
    "gradio>=6.1.0",
    "langchain>=1.1.3",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-google-genai>=4.0.0",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "lxml>=6.0.2",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "wikipedia>=1.4.0",
]
